+++ indexes/books.js
@@ -5,7 +5,7 @@ const books = await readdir('./indexes/json/books');
 
 const compositeJson = await Promise.all(books.map(async (fileName) => {
   return {
-    book: fileName.replace(/\.json$/gu, ''),
+    book: fileName.replaceAll(/\.json$/gu, ''),
     index: JSON.parse(
       await readFile(
         new URL(`json/books/${fileName}`, import.meta.url)